+++ src/live_executor.cc
@@ -5,8 +5,11 @@
 #include "grid.h"
 #include "host_object.h"
 #include "instruction_graph.h"
+#include "log.h"
 #include "named_threads.h"
 #include "out_of_order_engine.h"
+#include "print_utils.h"
+#include "print_utils_internal.h"
 #include "receive_arbiter.h"
 #include "system_info.h"
 #include "tracy.h"
