+++ test/device_selection_tests.cc
@@ -8,7 +8,7 @@
 #include <catch2/generators/catch_generators.hpp>
 #include <catch2/matchers/catch_matchers_string.hpp>
 
-#include "device_selection.h"
+#include "select_devices.h"
 #include "test_utils.h"
 
 using dt = sycl::info::device_type;
@@ -137,44 +137,44 @@ TEST_CASE("check_required_device_aspects throws if a device does not support req
 	CHECK_THROWS_WITH(check_required_device_aspects(device), "device does not support USM device allocations");
 }
 
-TEST_CASE("pick_devices prefers user-specified device list", "[device-selection]") {
+TEST_CASE("select_devices prefers user-specified device list", "[device-selection]") {
 	test_utils::allow_max_log_level(log_level::warn);
 
 	const host_config h_cfg{1, 0};
 	auto [mp] = create_mock_platforms(std::nullopt);
 
-	CHECK_THROWS_WITH(
-	    pick_devices(h_cfg, std::vector<mock_device>{}, std::vector<mock_platform>{mp}), "Device selection failed: The user-provided list of devices is empty");
+	CHECK_THROWS_WITH(select_devices(h_cfg, std::vector<mock_device>{}, std::vector<mock_platform>{mp}),
+	    "Device selection failed: The user-provided list of devices is empty");
 
 	const auto devices = mp.create_devices(dt::gpu, dt::gpu, dt::cpu, dt::accelerator);
-	const auto selected = pick_devices(h_cfg, std::vector<mock_device>{devices.begin(), devices.end()}, std::vector<mock_platform>{mp});
+	const auto selected = select_devices(h_cfg, std::vector<mock_device>{devices.begin(), devices.end()}, std::vector<mock_platform>{mp});
 	CHECK(selected == std::vector<mock_device>{devices.begin(), devices.end()});
 }
 
-TEST_CASE("pick_devices requires user-specified devices to have the same backend", "[device-selection]") {
+TEST_CASE("select_devices requires user-specified devices to have the same backend", "[device-selection]") {
 	const host_config h_cfg{1, 0};
 	auto [mp_1, mp_2] = create_mock_platforms(std::nullopt, std::nullopt);
 	auto [md_1] = mp_1.create_devices(dt::gpu);
 	auto [md_2] = mp_2.create_devices(dt::gpu);
 
-	CHECK_NOTHROW(pick_devices(h_cfg, std::vector<mock_device>{md_1, md_2}, std::vector<mock_platform>{mp_1, mp_2}));
+	CHECK_NOTHROW(select_devices(h_cfg, std::vector<mock_device>{md_1, md_2}, std::vector<mock_platform>{mp_1, mp_2}));
 
 	mp_1.set_backend(mock_device_backend::foo);
 	mp_2.set_backend(mock_device_backend::bar);
-	CHECK_THROWS_WITH(pick_devices(h_cfg, std::vector<mock_device>{md_1, md_2}, std::vector<mock_platform>{mp_1, mp_2}),
+	CHECK_THROWS_WITH(select_devices(h_cfg, std::vector<mock_device>{md_1, md_2}, std::vector<mock_platform>{mp_1, mp_2}),
 	    "Device selection failed: The user-provided list of devices contains devices from different backends");
 }
 
-TEST_CASE("pick_devices throws if a user-specified devices does not support required aspects", "[device-selection]") {
+TEST_CASE("select_devices throws if a user-specified devices does not support required aspects", "[device-selection]") {
 	const host_config h_cfg{1, 0};
 	auto [mp] = create_mock_platforms(std::nullopt);
 	auto [md_1, md_2] = mp.create_devices(dt::gpu, dt::gpu);
 	md_2.set_aspect(sycl::aspect::usm_device_allocations, false);
-	CHECK_THROWS_WITH(pick_devices(h_cfg, std::vector<mock_device>{md_1, md_2}, std::vector<mock_platform>{mp}),
+	CHECK_THROWS_WITH(select_devices(h_cfg, std::vector<mock_device>{md_1, md_2}, std::vector<mock_platform>{mp}),
 	    "Device selection failed: Device 1 in user-provided list of devices caused error: device does not support USM device allocations");
 }
 
-TEST_CASE("pick_devices selects the largest subset of GPUs that share the same backend", "[device-selection]") {
+TEST_CASE("select_devices selects the largest subset of GPUs that share the same backend", "[device-selection]") {
 	const host_config h_cfg{1, 0};
 	auto [mp_1, mp_2, mp_3] = create_mock_platforms(std::nullopt, std::nullopt, std::nullopt);
 	mp_1.create_devices(dt::gpu);
@@ -185,11 +185,11 @@ TEST_CASE("pick_devices selects the largest subset of GPUs that share the same b
 	mp_2.set_backend(mock_device_backend::bar);
 	mp_3.set_backend(mock_device_backend::qux);
 
-	const auto selected = pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp_1, mp_2, mp_3});
+	const auto selected = select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp_1, mp_2, mp_3});
 	CHECK(selected == std::vector<mock_device>{md_2, md_3, md_4});
 }
 
-TEST_CASE("pick_devices falls back to other device types if no GPUs are available", "[device-selection]") {
+TEST_CASE("select_devices falls back to other device types if no GPUs are available", "[device-selection]") {
 	test_utils::allow_max_log_level(log_level::warn);
 
 	const host_config h_cfg{1, 0};
@@ -202,16 +202,16 @@ TEST_CASE("pick_devices falls back to other device types if no GPUs are availabl
 	mp_2.set_backend(mock_device_backend::bar);
 	mp_3.set_backend(mock_device_backend::qux);
 
-	const auto selected_1 = pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp_1, mp_2, mp_3});
+	const auto selected_1 = select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp_1, mp_2, mp_3});
 	CHECK(selected_1 == std::vector<mock_device>{md_2, md_3, md_4});
 
 	// Once there is a GPU however, it takes precedence
 	auto [md_9] = mp_3.create_devices(dt::gpu);
-	const auto selected_2 = pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp_1, mp_2, mp_3});
+	const auto selected_2 = select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp_1, mp_2, mp_3});
 	CHECK(selected_2 == std::vector<mock_device>{md_9});
 }
 
-TEST_CASE("pick_devices only considers devices that support required aspects", "[device-selection]") {
+TEST_CASE("select_devices only considers devices that support required aspects", "[device-selection]") {
 	const host_config h_cfg{1, 0};
 	auto type = GENERATE(dt::gpu, dt::cpu);
 	auto [mp_1, mp_2, mp_3] = create_mock_platforms(std::nullopt, std::nullopt, std::nullopt);
@@ -226,11 +226,11 @@ TEST_CASE("pick_devices only considers devices that support required aspects", "
 	md_2.set_aspect(sycl::aspect::usm_device_allocations, false);
 	md_3.set_aspect(sycl::aspect::usm_device_allocations, false);
 
-	const auto selected = pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp_1, mp_2, mp_3});
+	const auto selected = select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp_1, mp_2, mp_3});
 	CHECK(selected == std::vector<mock_device>{md_5, md_6});
 }
 
-TEST_CASE("pick_devices supports passing a device selector function", "[device-selection]") {
+TEST_CASE("select_devices supports passing a device selector function", "[device-selection]") {
 	const host_config h_cfg{1, 0};
 	auto [mp_1, mp_2] = create_mock_platforms(std::nullopt, std::nullopt);
 	auto [md_1] = mp_1.create_devices(dt::gpu);
@@ -244,11 +244,11 @@ TEST_CASE("pick_devices supports passing a device selector function", "[device-s
 	const auto to_select = std::array<mock_device, 2>{md_1, md_2}[device_idx];
 
 	auto device_selector = [to_select](const mock_device& d) { return d == to_select ? 2 : 1; };
-	auto selected = pick_devices(h_cfg, device_selector, std::vector<mock_platform>{mp_1, mp_2});
+	auto selected = select_devices(h_cfg, device_selector, std::vector<mock_platform>{mp_1, mp_2});
 	CHECK(selected == std::vector<mock_device>{to_select});
 }
 
-TEST_CASE("pick_devices selects the subset of devices with the largest cumulative selector score sharing the same backend", "[device-selection]") {
+TEST_CASE("select_devices selects the subset of devices with the largest cumulative selector score sharing the same backend", "[device-selection]") {
 	const host_config h_cfg{1, 0};
 	auto [mp_1, mp_2, mp_3] = create_mock_platforms(std::nullopt, std::nullopt, std::nullopt);
 	auto [md_1] = mp_1.create_devices(dt::gpu);
@@ -268,7 +268,7 @@ TEST_CASE("pick_devices selects the subset of devices with the largest cumulativ
 	    {/* md_1 */ 50}, {/* md_2 */ 10, /* md_3 */ 20, /* md_4 */ ignore_md_4 ? -1 : 30}, {/* md_5 */ 30, /* md_6 */ 40}};
 	const auto selector = [&scores](const mock_device& d) { return scores[d.get_platform().get_id()][d.get_id()]; };
 
-	const auto selected = pick_devices(h_cfg, selector, std::vector<mock_platform>{mp_1, mp_2, mp_3});
+	const auto selected = select_devices(h_cfg, selector, std::vector<mock_platform>{mp_1, mp_2, mp_3});
 
 	if(ignore_md_4) {
 		if(md_6_no_usm) {
@@ -285,43 +285,43 @@ TEST_CASE("pick_devices selects the subset of devices with the largest cumulativ
 	}
 }
 
-TEST_CASE("pick_devices does not consider devices with a negative selector score", "[device-selection]") {
+TEST_CASE("select_devices does not consider devices with a negative selector score", "[device-selection]") {
 	const host_config h_cfg{1, 0};
 	auto [mp] = create_mock_platforms(std::nullopt);
 	auto [md_1, md_2, md_3] = mp.create_devices(dt::gpu, dt::gpu, dt::gpu);
 
 	auto selector = [md_2 = md_2](const mock_device& d) { return d.get_id() == md_2.get_id() ? -1 : 1; };
-	const auto selected = pick_devices(h_cfg, selector, std::vector<mock_platform>{mp});
+	const auto selected = select_devices(h_cfg, selector, std::vector<mock_platform>{mp});
 	CHECK(selected == std::vector<mock_device>{md_1, md_3});
 }
 
-TEST_CASE("pick_devices throws if no devices are available", "[device-selection]") {
+TEST_CASE("select_devices throws if no devices are available", "[device-selection]") {
 	const host_config h_cfg{1, 0};
 	auto [mp] = create_mock_platforms(std::nullopt);
 
 	SECTION("from the start") {
 		const auto selector = [](const mock_device&) { return -1; };
-		CHECK_THROWS_WITH(pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp}), "Device selection failed: No devices available");
-		CHECK_THROWS_WITH(pick_devices(h_cfg, selector, std::vector<mock_platform>{mp}), "Device selection failed: No devices available");
+		CHECK_THROWS_WITH(select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp}), "Device selection failed: No devices available");
+		CHECK_THROWS_WITH(select_devices(h_cfg, selector, std::vector<mock_platform>{mp}), "Device selection failed: No devices available");
 	}
 
 	SECTION("if all are ignored due to missing aspects") {
 		auto [md_1] = mp.create_devices(dt::gpu);
 		md_1.set_aspect(sycl::aspect::usm_device_allocations, false);
-		CHECK_THROWS_WITH(pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp}), "Device selection failed: No eligible devices found");
+		CHECK_THROWS_WITH(select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp}), "Device selection failed: No eligible devices found");
 		auto [md_2] = mp.create_devices(dt::cpu);
 		md_2.set_aspect(sycl::aspect::usm_device_allocations, false);
-		CHECK_THROWS_WITH(pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp}), "Device selection failed: No eligible devices found");
+		CHECK_THROWS_WITH(select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp}), "Device selection failed: No eligible devices found");
 	}
 
 	SECTION("if all are discarded by selector") {
 		auto [md_1] = mp.create_devices(dt::gpu);
 		const auto selector = [](const mock_device&) { return -1; };
-		CHECK_THROWS_WITH(pick_devices(h_cfg, selector, std::vector<mock_platform>{mp}), "Device selection failed: No eligible devices found");
+		CHECK_THROWS_WITH(select_devices(h_cfg, selector, std::vector<mock_platform>{mp}), "Device selection failed: No eligible devices found");
 	}
 }
 
-TEST_CASE("pick_devices attempts to evenly distributed devices if there is more than one local node", "[device-selection]") {
+TEST_CASE("select_devices attempts to evenly distributed devices if there is more than one local node", "[device-selection]") {
 	struct distribution {
 		size_t num_devices;
 		std::vector<size_t> devices_per_node;
@@ -349,7 +349,7 @@ TEST_CASE("pick_devices attempts to evenly distributed devices if there is more 
 		std::unordered_set<std::pair<size_t, size_t>, utils::pair_hash> unique_devices;
 		for(size_t i = 0; i < dist.devices_per_node.size(); ++i) {
 			const host_config h_cfg{dist.devices_per_node.size(), i};
-			const auto selected = pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp});
+			const auto selected = select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp});
 			REQUIRE_LOOP(selected.size() == dist.devices_per_node[i]);
 			for(const auto& d : selected) {
 				const auto unique_id = std::pair{d.get_platform().get_id(), d.get_id()};
@@ -363,7 +363,7 @@ TEST_CASE("pick_devices attempts to evenly distributed devices if there is more 
 	                                                      "attempt to distribute local devices evenly across nodes."));
 }
 
-TEST_CASE("pick_devices distributes devices in round-robin fashion if there are fewer than local nodes", "[device-selection]") {
+TEST_CASE("select_devices distributes devices in round-robin fashion if there are fewer than local nodes", "[device-selection]") {
 	auto [mp] = create_mock_platforms(std::nullopt);
 	auto devices = mp.create_devices(dt::gpu, dt::gpu, dt::gpu);
 
@@ -373,7 +373,7 @@ TEST_CASE("pick_devices distributes devices in round-robin fashion if there are 
 
 	for(size_t i = 0; i < num_ranks; ++i) {
 		const host_config h_cfg{num_ranks, i};
-		const auto selected = pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp});
+		const auto selected = select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp});
 		REQUIRE(selected.size() == 1);
 		CHECK(selected[0] == devices[i % 3]);
 		CHECK(test_utils::log_contains_exact(
@@ -381,26 +381,26 @@ TEST_CASE("pick_devices distributes devices in round-robin fashion if there are 
 	}
 }
 
-TEST_CASE("pick_devices prints device and platform information", "[device-selection]") {
+TEST_CASE("select_devices prints device and platform information", "[device-selection]") {
 	const host_config h_cfg{1, 0};
 	mock_platform mp(68, "My platform");
 	auto mds = mp.create_devices(type_and_name{dt::gpu, "My first device"}, type_and_name{dt::gpu, "My second device"});
 
 	SECTION("when devices are provided by user") {
-		pick_devices(h_cfg, std::vector<mock_device>{mds[0], mds[1]}, std::vector<mock_platform>{mp});
+		select_devices(h_cfg, std::vector<mock_device>{mds[0], mds[1]}, std::vector<mock_platform>{mp});
 		CHECK(test_utils::log_contains_exact(log_level::info, "Using platform \"My platform\", device \"My first device\" as D0 (specified by user)"));
 		CHECK(test_utils::log_contains_exact(log_level::info, "Using platform \"My platform\", device \"My second device\" as D1 (specified by user)"));
 	}
 
 	SECTION("when automatically selecting a device") {
-		pick_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp});
+		select_devices(h_cfg, auto_select_devices{}, std::vector<mock_platform>{mp});
 		CHECK(test_utils::log_contains_exact(log_level::info, "Using platform \"My platform\", device \"My first device\" as D0 (automatically selected)"));
 		CHECK(test_utils::log_contains_exact(log_level::info, "Using platform \"My platform\", device \"My second device\" as D1 (automatically selected)"));
 	}
 
 	SECTION("when a device selector is provided") {
 		const auto selector = [mds](const mock_device&) { return 100; };
-		pick_devices(h_cfg, selector, std::vector<mock_platform>{mp});
+		select_devices(h_cfg, selector, std::vector<mock_platform>{mp});
 		CHECK(test_utils::log_contains_exact(log_level::info, "Using platform \"My platform\", device \"My first device\" as D0 (via user-provided selector)"));
 		CHECK(
 		    test_utils::log_contains_exact(log_level::info, "Using platform \"My platform\", device \"My second device\" as D1 (via user-provided selector)"));
